# winalog-go/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/query": {
            "post": {
                "description": "初始化QueryHandler",
                "tags": [
                    "query"
                ],
                "summary": "创建查询处理器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据库实例",
                        "name": "db",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/query/quick-queries": {
            "get": {
                "description": "返回所有预定义的快速查询语句",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "列出快速查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按分类过滤",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.QuickQuery"
                            }
                        }
                    }
                }
            }
        },
        "/api/query/quick-query/{name}": {
            "get": {
                "description": "返回指定快速查询的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "获取快速查询详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.QuickQuery"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/query/tables": {
            "get": {
                "description": "返回数据库中所有可查询的表及其结构信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "列出数据表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.TableInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/system/tasks": {
            "get": {
                "description": "返回系统计划任务列表(Windows)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取计划任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "是否启用",
                        "name": "enabled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TaskResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/system/users": {
            "get": {
                "description": "返回系统本地用户列表(Windows)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "是否启用",
                        "name": "enabled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ColumnInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "not_null": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.ErrorCode"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.QuickQuery": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sql": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
...
        "api.TableInfo": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ColumnInfo"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "api.TaskInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "api.TaskResponse": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.TaskInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.UserInfo": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "home_dir": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.UserInfo"
                    }
                }
            }
        },
...
        "types.ErrorCode": {
            "type": "string",
            "enum": [
                "SUCCESS",
                "INTERNAL_ERROR",
                "INVALID_PARAM",
                "INVALID_REQUEST",
                "INVALID_QUERY",
                "NOT_FOUND",
                "EVENT_NOT_FOUND",
                "ALERT_NOT_FOUND",
                "ALERT_ALREADY_RESOLVED",
                "UNAUTHORIZED",
                "PARSE_FAILED",
                "IMPORT_FAILED",
                "FILE_NOT_FOUND",
                "FILE_LOCKED",
                "INVALID_FORMAT",
                "DB_ERROR",
                "DB_READ_ONLY",
                "RULE_INVALID",
                "RULE_NOT_FOUND",
                "RULE_DISABLED",
                "SEARCH_FAILED",
                "RESULT_TOO_LARGE",
                "HASH_MISMATCH",
                "SIGNATURE_INVALID",
                "NOT_SUPPORTED"
            ],
            "x-enum-varnames": [
                "ErrCodeSuccess",
                "ErrCodeInternalError",
                "ErrCodeInvalidParam",
                "ErrCodeInvalidRequest",
                "ErrCodeInvalidQuery",
                "ErrCodeNotFound",
                "ErrCodeEventNotFound",
                "ErrCodeAlertNotFound",
                "ErrCodeAlertAlreadyResolved",
                "ErrCodeUnauthorized",
                "ErrCodeParseFailed",
                "ErrCodeImportFailed",
                "ErrCodeFileNotFound",
                "ErrCodeFileLocked",
                "ErrCodeInvalidFormat",
                "ErrCodeDBError",
                "ErrCodeDBReadOnly",
                "ErrCodeRuleInvalid",
                "ErrCodeRuleNotFound",
                "ErrCodeRuleDisabled",
                "ErrCodeSearchFailed",
                "ErrCodeResultTooLarge",
                "ErrCodeHashMismatch",
                "ErrCodeSignatureInvalid",
                "ErrCodeNotSupported"
            ]
        },
...
    }
}